+++ pkg/vm/hash-internal.go
@@ -39,8 +39,10 @@ func (h *Hash) Set(key, value *Value) error {
 		h.internalMap[createHashString(key.GetInt64())] = value
 	case FloatId:
 		h.internalMap[createHashString(key.GetFloat64())] = value
+	default:
+		return NotHashable
 	}
-	return NotHashable
+	return nil
 }
 
 func (h *Hash) Get(key *Value) (*Value, error) {
@@ -75,8 +77,44 @@ func (h *Hash) Del(key *Value) error {
 		delete(h.internalMap, createHashString(key.GetInt64()))
 	case FloatId:
 		delete(h.internalMap, createHashString(key.GetFloat64()))
+	default:
+		return NotHashable
 	}
-	return NotHashable
+	return nil
+}
+
+func (h *Hash) Copy() *Hash {
+	h.mutex.Lock()
+	defer h.mutex.Unlock()
+	result := &Hash{
+		mutex:       &sync.Mutex{},
+		internalMap: make(map[string]*Value, len(h.internalMap)),
+	}
+	for key, value := range h.internalMap {
+		result.internalMap[key] = value
+	}
+	return result
+}
+
+func (h *Hash) In(key *Value) (bool, error) {
+	h.mutex.Lock()
+	defer h.mutex.Unlock()
+	var found bool
+	switch key.TypeId() {
+	case StringId:
+		_, found = h.internalMap[createHashString(string(key.GetBytes()))]
+	case BytesId:
+		_, found = h.internalMap[createHashString(key.GetBytes())]
+	case BoolId:
+		_, found = h.internalMap[createHashString(key.GetBool())]
+	case IntId:
+		_, found = h.internalMap[createHashString(key.GetInt64())]
+	case FloatId:
+		_, found = h.internalMap[createHashString(key.GetFloat64())]
+	default:
+		return false, NotHashable
+	}
+	return found, nil
 }
 
 func (plasma *Plasma) NewInternalHash() *Hash {