+++ pkg/vm/hash.go
@@ -1,5 +1,7 @@
 package vm
 
+import magic_functions "github.com/shoriwe/gplasma/pkg/common/magic-functions"
+
 func (plasma *Plasma) hashClass() *Value {
 	class := plasma.NewValue(plasma.rootSymbols, BuiltInClassId, plasma.class)
 	class.SetAny(func(argument ...*Value) (*Value, error) {
@@ -10,25 +12,73 @@ func (plasma *Plasma) hashClass() *Value {
 
 /*
 NewHash magic function:
-TODO In                  __in__
-TODO Equals              __equals__
-TODO NotEqual            __not_equal__
-TODO Mul                 __mul__
-TODO Length              __len__
-TODO Bool                __bool__
-TODO String              __string__
-TODO Bytes               __bytes__
-TODO Array               __array__
-TODO Tuple               __tuple__
-TODO Get                 __get__
-TODO Set                 __set__
-TODO Del				 __del__
-TODO Copy                __copy__
-TODO Iter                __iter__
+In                  __in__
+Length              __len__
+Bool                __bool__
+String              __string__
+Bytes               __bytes__
+Get                 __get__
+Set                 __set__
+Del				 	__del__
+Copy                __copy__
 */
 func (plasma *Plasma) NewHash(hash *Hash) *Value {
 	result := plasma.NewValue(plasma.rootSymbols, HashId, plasma.hash)
 	result.SetAny(hash)
-	// TODO: init magic functions
+	result.Set(magic_functions.In, plasma.NewBuiltInFunction(
+		result.vtable,
+		func(argument ...*Value) (*Value, error) {
+			in, inError := result.GetHash().In(argument[0])
+			return plasma.NewBool(in), inError
+		},
+	))
+	result.Set(magic_functions.Length, plasma.NewBuiltInFunction(
+		result.vtable,
+		func(argument ...*Value) (*Value, error) {
+			return plasma.NewInt(result.GetHash().Size()), nil
+		},
+	))
+	result.Set(magic_functions.Bool, plasma.NewBuiltInFunction(
+		result.vtable,
+		func(argument ...*Value) (*Value, error) {
+			return plasma.NewBool(result.Bool()), nil
+		},
+	))
+	result.Set(magic_functions.String, plasma.NewBuiltInFunction(
+		result.vtable,
+		func(argument ...*Value) (*Value, error) {
+			return plasma.NewString([]byte(result.String())), nil
+		},
+	))
+	result.Set(magic_functions.Bytes, plasma.NewBuiltInFunction(
+		result.vtable,
+		func(argument ...*Value) (*Value, error) {
+			return plasma.NewBytes([]byte(result.String())), nil
+		},
+	))
+	result.Set(magic_functions.Get, plasma.NewBuiltInFunction(
+		result.vtable,
+		func(argument ...*Value) (*Value, error) {
+			return result.GetHash().Get(argument[0])
+		},
+	))
+	result.Set(magic_functions.Set, plasma.NewBuiltInFunction(
+		result.vtable,
+		func(argument ...*Value) (*Value, error) {
+			return plasma.none, result.GetHash().Set(argument[0], argument[1])
+		},
+	))
+	result.Set(magic_functions.Del, plasma.NewBuiltInFunction(
+		result.vtable,
+		func(argument ...*Value) (*Value, error) {
+			return plasma.none, result.GetHash().Del(argument[0])
+		},
+	))
+	result.Set(magic_functions.Copy, plasma.NewBuiltInFunction(
+		result.vtable,
+		func(argument ...*Value) (*Value, error) {
+			return plasma.NewHash(result.GetHash().Copy()), nil
+		},
+	))
 	return result
 }